4.machine_learning/SlackBot/new/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
Flask==1.1.4
html5lib==1.1
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
multidict==5.1.0
pyee==7.0.4
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
slack==0.0.2
slackclient==2.9.3
slackeventsapi==2.2.1
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.6.3
